+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/authentication/signin/SignInViewModel.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
-class SignInViewModel constructor(
+class SignInViewModel(
     private val signInUseCase: SignInUseCase
 ) : ViewModel() {
