+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/cart/CartFragment.kt
@@ -9,8 +9,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-
-
 class CartFragment : BottomSheetDialogFragment() {
 
     override fun onCreateView(
