+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/checkout/CheckOutViewModel.kt
@@ -18,7 +18,7 @@ import kotlinx.coroutines.launch
 
 
 
-class CheckOutViewModel constructor(
+class CheckOutViewModel(
     private val getCurrentCartUseCase: GetCurrentCartUseCase,
     private val getProfileUseCase: GetProfileUseCase,
     private val submitOrderUseCase: SubmitOrderUseCase
