+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/mealplanning/overview/OverviewViewModel.kt
@@ -15,9 +15,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
-
-
-class OverviewViewModel constructor(
+class OverviewViewModel(
     private val addMealToPlanUseCase: AddMealToPlanUseCase,
     private val retrieveMealByTypeUseCase: RetrieveMealByTypeUseCase,
     private val removeMealFromPlanUseCase: RemoveMealFromPlanUseCase
