+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/orderhistory/OrderHistoryFragment.kt
@@ -7,8 +7,6 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 
-
-
 class OrderHistoryFragment : Fragment() {
 
     override fun onCreateView(
