+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/productdetail/ProductDetailFragment.kt
@@ -8,8 +8,6 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 
-
-
 class ProductDetailFragment : Fragment() {
 
     override fun onCreateView(
