+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/productdetail/ProductDetailViewModel.kt
@@ -16,7 +16,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 
-class ProductDetailViewModel constructor(
+class ProductDetailViewModel(
     savedStateHandle: SavedStateHandle,
     getProductDetailUseCase: GetProductDetailUseCase,
     private val orderRepository: OrderRepository
